requirements.txt
flask==2.0.2
feedgen==0.9.0
google-cloud-datastore==2.4.0
